requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.1
lxml==4.9.3
jinja2==3.1.2
python-multipart==0.0.6
google-generativeai==0.3.1
python-dotenv==1.0.0 